requirements.txt
click
ffmpeg
matplotlib
numpy
opencv-python
pillow
scikit-image
termcolor
tqdm
